Refactor manage functions into setup
