+++ manage.py
@@ -1,29 +1,6 @@
-import connexion
-from utils import combine_specifications
-from firebase_admin import credentials, initialize_app
+from setup import initialize_firebase, start_server, FIREBASE_CERT_PATH
 
-
-# Initializing Firestore database
-# Can import firestore_db to utilize database
-def initialize_firebase():
-    initialize_app(credentials.Certificate("serviceAccountKey.json"))
-
-
-def start_server(name: str):
-    app = connexion.App(name)
-    # Tell app to read the combined openapi specification
-    app.add_api(
-        combine_specifications(
-            "./base_swagger.yaml",
-            "./app/poi_api/spec.yaml",
-            "./app/user/spec.yaml",
-            "./app/rewards/spec.yaml",
-        )
-    )
-    return app
-
-
-initialize_firebase()
+initialize_firebase(FIREBASE_CERT_PATH)
 app = start_server(__name__)
